requirements.txt
discord.py
python-dotenv
pytz
flask
pillow
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib>=0.5.2
google-auth